+++ src/test/java/com/tregouet/tree_finder/algo/unidimensional_sorting/impl/UnidimensionalSorterTest.java
@@ -1,6 +1,6 @@
 package com.tregouet.tree_finder.algo.unidimensional_sorting.impl;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -10,187 +10,97 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.jgrapht.Graphs;
-import org.jgrapht.alg.TransitiveReduction;
-import org.jgrapht.graph.DefaultEdge;
 import org.jgrapht.graph.DirectedAcyclicGraph;
 import org.jgrapht.traverse.TopologicalOrderIterator;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.google.common.collect.Sets;
 import com.tregouet.tree_finder.algo.unidimensional_sorting.IUnidimensionalSorter;
+import com.tregouet.tree_finder.algo.unidimensional_sorting.impl.UnidimensionalSorter;
 import com.tregouet.tree_finder.data.Tree;
 import com.tregouet.tree_finder.data.UpperSemilattice;
 import com.tregouet.tree_finder.error.InvalidInputException;
 
+import utils.DichotomizableString;
 import utils.EdgeForTests;
-import utils.RandomUSLGenerator;
+import utils.Visualizer;
 
-@SuppressWarnings("unused")
 public class UnidimensionalSorterTest {
 	
-	private final String a = "A";
-	private final String b = "B";
-	private final String c = "C";
-	private final String d = "D";
-	private final Set<String> atoms = new HashSet<>(Arrays.asList(new String[] {a, b, c, d}));
-	//dataset1
-	private UpperSemilattice<Set<String>, EdgeForTests> powerSetUSL = null;
-	//dataset2
-	private UpperSemilattice<String, EdgeForTests> notComplementedUSL = null;
+	private final DichotomizableString a = new DichotomizableString("A");
+	private final DichotomizableString b = new DichotomizableString("B");
+	private final DichotomizableString c = new DichotomizableString("C");
+	private final DichotomizableString d = new DichotomizableString("D");
+	private final DichotomizableString e = new DichotomizableString("E");
+	private UpperSemilattice<DichotomizableString, EdgeForTests> notComplementedUSL = null;
 
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
 	}
 
 	@Before
-	public void setUp() throws Exception {
+	public void setUp() {
 	}
-	
+
 	@Test
-	public void whenSortingsOfPowerSetUSLRequestedThenReturned() throws InvalidInputException, IOException {
-		setUpPowerSet();
-		IUnidimensionalSorter<Set<String>, EdgeForTests> sorter = 
-				new UnidimensionalSorter<Set<String>, EdgeForTests>(powerSetUSL);
-		Collection<Tree<Set<String>, EdgeForTests>> sortingTrees = sorter.getSortingTrees();
-		assertTrue(sortingTrees.size() > 0);
-	}
-	
-	
-	@Test
-	public void whenSortingsOfNotComplementedUSLRequestedThenReturned() throws InvalidInputException, IOException {
+	public void whenSortingsOfNotComplementedUSLRequestedThenReturned() throws IOException, InvalidInputException {
 		setUpNotComplementedUSL();
 		/*
-		Visualizer.visualize(((DirectedAcyclicGraph<String, EdgeForTests>) notComplementedUSL), "2111030946_USL", 'a');
+		Visualizer.visualize(
+				((DirectedAcyclicGraph<DichotomizableString, EdgeForTests>) notComplementedUSL), 
+				"2111030946_USL", 0.0);
 		*/
-		IUnidimensionalSorter<String, EdgeForTests> sorter = 
-				new UnidimensionalSorter<String, EdgeForTests>(notComplementedUSL);
-		Collection<Tree<String, EdgeForTests>> sortingTrees = sorter.getSortingTrees();
+		IUnidimensionalSorter<DichotomizableString, EdgeForTests> sorter = 
+				new UnidimensionalSorter<DichotomizableString, EdgeForTests>(notComplementedUSL);
+		Collection<Tree<DichotomizableString, EdgeForTests>> sortingTrees = sorter.getSortingTrees();
 		/*
 		int treeIdx = 0;
-		for (Tree<String, EdgeForTests> tree : sortingTrees){
-			Visualizer.visualize(tree, "2111030946_RI_tree" + Integer.toString(treeIdx++), 'a');
+		for (Tree<DichotomizableString, EdgeForTests> tree : sortingTrees){
+			Visualizer.visualize(tree, "2111261532_D_tree" + Integer.toString(treeIdx++), 0.0);
 		}
 		*/
 		assertTrue(sortingTrees.size() > 0);
-	}
-	
-	@Test
-	public void whenSortingOfRandomUSLRequestedThenReturned() throws InvalidInputException, IOException {
-		boolean returned = true;
-		int maxNbOfTrials = 1;
-		int maxNbOfLeaves = 5;
-		int nbOfTreesReturned = 0;
-		RandomUSLGenerator generator = new RandomUSLGenerator(maxNbOfLeaves);
-		for (int i = 0 ; i < maxNbOfTrials ; i++) {
-			UpperSemilattice<Set<String>, DefaultEdge> randomUSL = generator.nextRandomUSL();
-			TransitiveReduction.INSTANCE.reduce(randomUSL);
-			/*
-			Visualizer.visualize(randomUSL, "2111011144_usl" + i, 0);
-			*/
-			nbOfTreesReturned = 0;
-			IUnidimensionalSorter<Set<String>, DefaultEdge> sorter = new UnidimensionalSorter<>(randomUSL);
-			while (sorter.hasNext()) {
-				Tree<Set<String>, DefaultEdge> nextTree = sorter.next();
-				/*
-				Visualizer.visualize(nextTree, "2111011144_usl" + i + "tree" + nbOfTreesReturned, 0);
-				*/
-				nbOfTreesReturned++;
-			}
-			if (nbOfTreesReturned == 0)
-				returned = false;
-			nbOfTreesReturned = 0;
-		}
-		assertTrue(returned);
-	}
-	
-	@Test
-	public void whenTopologicalOrderRequestedThenContainsEveryElement() throws InvalidInputException {
-		setUpPowerSet();
-		boolean everyElementFound = true;
-		/*
-		Visualizer.visualize(((DirectedAcyclicGraph<String, EdgeForTests>) notComplementedUSL), "2111030946_USL", 'a');
-		*/
-		IUnidimensionalSorter<Set<String>, EdgeForTests> sorter = 
-				new UnidimensionalSorter<Set<String>, EdgeForTests>(powerSetUSL);
-		Collection<Tree<Set<String>, EdgeForTests>> sortingTrees = sorter.getSortingTrees();
-		/*
-		int treeIdx = 0;
-		for (Tree<String, EdgeForTests> tree : sortingTrees){
-			Visualizer.visualize(tree, "2111030946_RI_tree" + Integer.toString(treeIdx++), 'a');
-		}
-		*/
-		for (Tree<Set<String>, EdgeForTests> sortingTree : sortingTrees) {
-			if (!sortingTree.vertexSet().equals(new HashSet<Set<String>>(sortingTree.getTopologicalOrder())))
-				everyElementFound = false;
-		}
-		assertTrue(sortingTrees.size() > 0);
-	}
-	
-	private void setUpPowerSet() {
-		DirectedAcyclicGraph<Set<String>, EdgeForTests> upperSemiLattice = new DirectedAcyclicGraph<>(null, EdgeForTests::new, false);
-		List<Set<String>> vertices = new ArrayList<>(Sets.powerSet(atoms));
-		vertices.remove(new HashSet<String>());
-		Graphs.addAllVertices(upperSemiLattice, vertices);
-		for (int i = 0 ; i < vertices.size() - 1 ; i++) {
-			Set<String> iVertex = vertices.get(i);
-			for (int j = i + 1 ; j < vertices.size() ; j++) {
-				Set<String> jVertex = vertices.get(j);
-				if (iVertex.containsAll(jVertex))
-					upperSemiLattice.addEdge(jVertex, iVertex);
-				else if (jVertex.containsAll(iVertex))
-					upperSemiLattice.addEdge(iVertex, jVertex);
-			}
-		}
-		List<Set<String>> topoOrder = new ArrayList<>();
-		Set<Set<String>> leaves = new HashSet<>();
-		Set<String> root = null;;
-		TopologicalOrderIterator<Set<String>, EdgeForTests> topoIte = new TopologicalOrderIterator<>(upperSemiLattice);
-		while (topoIte.hasNext()) {
-			Set<String> nextElement = topoIte.next();
-			topoOrder.add(nextElement);
-			if (upperSemiLattice.inDegreeOf(nextElement) == 0)
-				leaves.add(nextElement);
-			if (upperSemiLattice.outDegreeOf(nextElement) == 0)
-				root = nextElement;
-		}
-		this.powerSetUSL = new UpperSemilattice<Set<String>, EdgeForTests>(upperSemiLattice, root, leaves, topoOrder);
 	}
 	
 	private void setUpNotComplementedUSL() {
-		String bc = "BC";
-		String cd = "CD";
-		String abc = "ABC";
-		String bcd = "BCD";
-		String abcd = "ABCD";
-		DirectedAcyclicGraph<String, EdgeForTests> notComplementedUSL = 
+		DichotomizableString ab = new DichotomizableString("AB");
+		DichotomizableString bc = new DichotomizableString("BC");
+		DichotomizableString cd = new DichotomizableString("CD");
+		DichotomizableString abc = new DichotomizableString("ABC");
+		DichotomizableString cde = new DichotomizableString("CDE");
+		DichotomizableString abcde = new DichotomizableString("ABCDE");
+		DirectedAcyclicGraph<DichotomizableString, EdgeForTests> notComplementedUSL = 
 				new DirectedAcyclicGraph<>(null, EdgeForTests::new, false);
 		notComplementedUSL.addVertex(a);
 		notComplementedUSL.addVertex(b);
 		notComplementedUSL.addVertex(c);
 		notComplementedUSL.addVertex(d);
+		notComplementedUSL.addVertex(e);
+		notComplementedUSL.addVertex(ab);
 		notComplementedUSL.addVertex(bc);
 		notComplementedUSL.addVertex(cd);
 		notComplementedUSL.addVertex(abc);
-		notComplementedUSL.addVertex(bcd);
-		notComplementedUSL.addVertex(abcd);
-		notComplementedUSL.addEdge(a, abc);
+		notComplementedUSL.addVertex(cde);
+		notComplementedUSL.addVertex(abcde);
+		notComplementedUSL.addEdge(a, ab);
+		notComplementedUSL.addEdge(b, ab);
 		notComplementedUSL.addEdge(b, bc);
 		notComplementedUSL.addEdge(c, bc);
 		notComplementedUSL.addEdge(c, cd);
 		notComplementedUSL.addEdge(d, cd);
+		notComplementedUSL.addEdge(e, cde);
+		notComplementedUSL.addEdge(ab, abc);
 		notComplementedUSL.addEdge(bc, abc);
-		notComplementedUSL.addEdge(bc, bcd);
-		notComplementedUSL.addEdge(cd, bcd);
-		notComplementedUSL.addEdge(abc, abcd);
-		notComplementedUSL.addEdge(bcd, abcd);
-		List<String> topologicalOrder =new ArrayList<>();
+		notComplementedUSL.addEdge(cd, cde);
+		notComplementedUSL.addEdge(abc, abcde);
+		notComplementedUSL.addEdge(cde, abcde);
+		List<DichotomizableString> topologicalOrder =new ArrayList<>();
 		new TopologicalOrderIterator<>(notComplementedUSL).forEachRemaining(topologicalOrder::add);
-		Set<String> leaves = new HashSet<>(Arrays.asList(new String[] {a, b, c, d}));
+		Set<DichotomizableString> leaves = new HashSet<>(Arrays.asList(new DichotomizableString[] {a, b, c, d, e}));
 		this.notComplementedUSL = 
-				new UpperSemilattice<String, EdgeForTests>(notComplementedUSL, abcd, leaves, topologicalOrder);
+				new UpperSemilattice<DichotomizableString, EdgeForTests>(
+						notComplementedUSL, abcde, leaves, topologicalOrder);
 	}
 
 }
